updated pr title and refactored components
