+++ components/VersionControlLinks.tsx
@@ -1,0 +1,20 @@
+export default function VersionControlLinks({
+  sha,
+  diffNum,
+}: {
+  sha: string;
+  diffNum: string | null;
+}) {
+  return (
+    <div>
+      <a href={`https://github.com/pytorch/pytorch/commit/${sha}`}>GitHub</a>
+      {diffNum !== undefined ? (
+        <span>
+          {" "}
+          |{" "}
+          <a href={`https://www.internalfb.com/diff/${diffNum}`}>Phabricator</a>
+        </span>
+      ) : null}
+    </div>
+  );
+}
