+++ lib/fetchPR.ts
@@ -6,7 +6,7 @@ export default async function fetchPR(pr: string): Promise<PRData[]> {
   const PRQuery = await rocksetClient.queryLambdas.executeQueryLambda(
     "commons",
     "prs_query",
-    "fb017667a2a06c9a",
+    "eea6932288753f6d",
     {
       parameters: [
         {
@@ -20,14 +20,14 @@ export default async function fetchPR(pr: string): Promise<PRData[]> {
   const shaSet = new Set();
   const results: PRData[] = [];
 
-  (PRQuery.results ?? []).map(({ sha, event_time, commit }) => {
+  (PRQuery.results ?? []).map(({ sha, event_time, commit, title }) => {
     if (!shaSet.has(sha)) {
       const shaString = sha as string;
       const eventTime = event_time as string;
       const firstLine = commit.indexOf("\n");
       const commitTitle: string = commit.slice(0, firstLine);
       shaSet.add(shaString);
-      results.push({ sha: shaString, eventTime, commitTitle });
+      results.push({ sha: shaString, eventTime, commitTitle, title });
     }
   });
   return results;
