+++ pages/commit/[sha].tsx
@@ -1,129 +1,8 @@
-import styles from "components/commit.module.css";
-import JobSummary from "components/JobSummary";
-import LogViewer from "components/LogViewer";
-import { isFailedJob } from "lib/jobUtils";
-import { CommitData, JobData } from "lib/types";
-import _ from "lodash";
+import CommitStatus from "components/CommitStatus";
 import { useRouter } from "next/router";
 import useSWR from "swr";
 
-function FilteredJobList({
-  filterName,
-  jobs,
-  pred,
-}: {
-  filterName: string;
-  jobs: JobData[];
-  pred: (job: JobData) => boolean;
-}) {
-  const filteredJobs = jobs.filter(pred);
-  if (filteredJobs.length === 0) {
-    return null;
-  }
-  return (
-    <div>
-      <h2>{filterName}</h2>
-      <ul>
-        {filteredJobs.map((job) => (
-          <li key={job.id}>
-            <JobSummary job={job} />
-            <LogViewer job={job} />
-          </li>
-        ))}
-      </ul>
-    </div>
-  );
-}
-
-function WorkflowBox({
-  workflowName,
-  jobs,
-}: {
-  workflowName: string;
-  jobs: JobData[];
-}) {
-  const isFailed = jobs.some(isFailedJob) !== false;
-  const workflowClass = isFailed
-    ? styles.workflowBoxFail
-    : styles.workflowBoxSuccess;
-  return (
-    <div className={workflowClass}>
-      <h3>{workflowName}</h3>
-      {jobs.map((job) => (
-        <div key={job.id}>
-          <JobSummary job={job} />
-          <LogViewer job={job} />
-        </div>
-      ))}
-    </div>
-  );
-}
-
-function WorkflowsContainer({ jobs }: { jobs: JobData[] }) {
-  const byWorkflow = _.groupBy(jobs, (job) => job.workflowName);
-  return (
-    <div className={styles.workflowContainer}>
-      {_.map(byWorkflow, (jobs, workflowName) => {
-        return (
-          <WorkflowBox
-            key={workflowName}
-            workflowName={workflowName}
-            jobs={jobs}
-          />
-        );
-      })}
-    </div>
-  );
-}
-
-function VersionControlLinks({
-  sha,
-  diffNum,
-}: {
-  sha: string;
-  diffNum: string | null;
-}) {
-  return (
-    <div>
-      <a href={`https://github.com/pytorch/pytorch/commit/${sha}`}>GitHub</a>
-      {diffNum !== undefined ? (
-        <span>
-          {" "}
-          |{" "}
-          <a href={`https://www.internalfb.com/diff/${diffNum}`}>Phabricator</a>
-        </span>
-      ) : null}
-    </div>
-  );
-}
-
 const fetcher = (url: string) => fetch(url).then((res) => res.json());
-
-export function CommitData({ commit }: { commit: CommitData }) {
-  return (
-    <>
-      <VersionControlLinks sha={commit.sha} diffNum={commit.diffNum} />
-
-      <article className={styles.commitMessage}>
-        {commit.commitMessageBody}
-      </article>
-
-      <FilteredJobList
-        filterName="Failed jobs"
-        jobs={commit.jobs}
-        pred={isFailedJob}
-      />
-
-      <FilteredJobList
-        filterName="Pending jobs"
-        jobs={commit.jobs}
-        pred={(job) => job.conclusion === "pending"}
-      />
-
-      <WorkflowsContainer jobs={commit.jobs} />
-    </>
-  );
-}
 
 export function CommitInfo({ sha }: { sha: string }) {
   const { data: commit, error } = useSWR(`/api/commit/${sha}`, fetcher, {
@@ -144,7 +23,7 @@ export function CommitInfo({ sha }: { sha: string }) {
   return (
     <div>
       <h2>{commit.commitTitle}</h2>
-      <CommitData commit={commit} />
+      <CommitStatus commit={commit} />
     </div>
   );
 }
